% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartcs-package.R
\docType{package}
\name{bartcs-package}
\alias{bartcs-package}
\title{bartcs: Bayesian Nonparametric Adjustment of Confounding}
\description{
Functions in \code{bartcs} serve one of three purposes.
\enumerate{
\item Functions for fitting: \code{sbart()}, \code{mbart()}.
\item Functions for summary: \code{summary()}, \code{gelman_rubin()}, \code{plot()}
\item Util function for openMP: \code{count_omp_thread()}
}
}
