% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_omp_thread}
\alias{count_omp_thread}
\title{Count the number of openMP threads for parallel computation}
\usage{
count_omp_thread()
}
\description{
\code{count_omp_thread()} counts the number of openMP threads for
parallel computation.
If it returns 1, openMP is not viable.
}
