% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelman_rubin.R
\name{gelman_rubin}
\alias{gelman_rubin}
\title{Gelman-Rubin diagnostic for \code{bartcs} objects.}
\usage{
gelman_rubin(x)
}
\arguments{
\item{x}{A \code{bartcs} object.}
}
\value{
Gelman-Rubin diagnostic value.
}
\description{
\code{gelman_rubin()} computes Gelman-Rubin diagnostic for \code{bartcs} objects.
}
