% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartcs-package.R
\docType{package}
\name{bartcs-package}
\alias{bartcs-package}
\title{bartcs: Bayesian Additive Regression Trees for Confounder Selection}
\description{
Fit Bayesian Regression Additive Trees (BART) models to
select true confounders from a large set of potential confounders and
to estimate average treatment effect. For more information, see
Kim et al. (2022) \url{doi:10.48550/arXiv.2203.11798}.
}
\details{
Functions in \code{bartcs} serve one of three purposes.
\enumerate{
\item Functions for fitting: \code{separate_bart()} and \code{single_bart()}.
\item Functions for summary: \code{summary()}, \code{plot()} and \code{gelman_rubin()}.
\item Utility function for OpenMP: \code{count_omp_thread()}.
}
}
\references{
Kim, C., Tec, M., & Zigler, C. M. (2022).
Bayesian Nonparametric Adjustment of Confounding.
\emph{arXiv preprint arXiv:2203.11798}.
\doi{10.48550/arXiv.2203.11798}
}
